backend/requirements.txt
flask==3.1.*
flask-cors==5.*
python-dotenv==1.*
requests==2.*
pyairtable==2.*
